requirements.txt
redis==3.5.3
pymongo==3.11.2
pyrsistent==0.17.3
pyTelegramBotAPI==4.14.0 
python-bitcoinrpc==1.0

web3==6.5.0

marshmallow==3.14.1
marshmallow-dataclass==8.5.3
marshmallow-enum==1.5.1
PyYAML==6.0
